Docker Deploy & Ini files + Database Files + Fixes Created the necessary configuration and initialisation files for deploying the web app with docker compose.
This includes the docker-compose.yml for docker compose configurations for
the web app, as well as configs and ini files for the nginx server,
and php server. The php server also includes a docker image for additional
setup of the php image's environment such as downloading the mysqli php package.

Also added sql files for mysql to use for creating the mymovielist database & tables,
as well as populating it with some starting data.

Additionally made some quality of life (QOL) fixes to the site, including
functionality, styling, and a security fix:
- Changed the servername for the database to "db" to match docker compose database container
for the connection in dbh.inc.php.
- Fixed the hover functionality of header's M.M.L. logo & the footer nav components
so that their hover styling's hover radius matches the text rather than extending out to the right.
- Fixed the javascript in makemovielist.php to clear the movie input textbox after a movie has been
added to the list so that the user doesn't need to clear the textbox themselves.
- Fixed the spacing between movie names so that there's now a command AND a space between movie names,
rather than just a comma.
- Added login error and login success messages to make the result of the login attempt clearer to the user.
- Made some styling fixes mostly to do with padding & margins as well as text size (header.php & makemovielist.php)
- Punction and typo fixes within some of the website's text.
- SECURITY FIX: INSERT SQL Query in makemovielist.inc.php inserted into the db
using the mysqli_query function and a query that referenced the data directly, making it vulnerable to SQL injection.
Fixed to use a prepare statement that binds the parameters afterwards instead. Also has the added benefit of
allowing more special characters in movie names due to better escaping of characters.