Email addresses made unique per user + other fixes Made email addresses unique per user by adding an additional check preventing signup if an entered email already exists in the database. Also updated the database creation queries accordingly so that the MySQL database enforces unique user emails using the UNIQUE constraint.
Also addressed a minor bug where an emptyfields error during signup would
also cause the emptyfields error message to show up next to the login form,
due to both of them checking for the same error parameter in the URL
and the login form also being present on the signup.php page in the header.

Added a little bit more detail in the readme on the containerization
of the web app, as well as other features the web app includes such as
database normalization and input validation.